+++ resolwe/flow/tests/test_api.py
@@ -471,6 +471,10 @@ class TestCollectionViewSetCase(TestCase):
     def setUp(self):
         super().setUp()
 
+        self.process = Process.objects.create(
+            name="Test process", contributor=self.contributor
+        )
+
         self.checkslug_viewset = CollectionViewSet.as_view(
             actions={"get": "slug_exists",}
         )
@@ -493,21 +497,25 @@ class TestCollectionViewSetCase(TestCase):
             actions={"get": "list", "post": "create",}
         )
 
+        self.list_url = reverse("resolwe-api:collection-list")
+
         self.detail_url = lambda pk: reverse(
             "resolwe-api:collection-detail", kwargs={"pk": pk}
         )
 
-    def _create_data(self):
-        process = Process.objects.create(
-            name="Test process", contributor=self.contributor,
+    def _create_data(self, data_status=None):
+        data = Data.objects.create(
+            name="Test data", contributor=self.contributor, process=self.process,
         )
 
-        return Data.objects.create(
-            name="Test data", contributor=self.contributor, process=process,
-        )
+        if data_status:
+            data.status = data_status
+            data.save()
+
+        return data
 
     def _create_entity(self):
-        return Entity.objects.create(name="Test entity", contributor=self.contributor,)
+        return Entity.objects.create(name="Test entity", contributor=self.contributor)
 
     def test_prefetch(self):
         descriptor_schema_1 = DescriptorSchema.objects.create(
@@ -695,6 +703,70 @@ class TestCollectionViewSetCase(TestCase):
         response = self.duplicate_viewset(request)
 
         self.assertEqual(response.data["detail"], MESSAGES["NOT_FOUND"])
+
+    def test_collection_data_count(self):
+        data1 = self._create_data()
+        data2 = self._create_data()
+
+        collection1 = Collection.objects.create(contributor=self.contributor)
+        assign_perm("view_collection", AnonymousUser(), collection1)
+
+        collection2 = Collection.objects.create(contributor=self.contributor)
+        assign_perm("view_collection", AnonymousUser(), collection2)
+        collection2.data.add(data1, data2)
+
+        response = self.client.get(self.detail_url(collection1.id))
+        self.assertEqual(response.data['data_count'], 0)
+
+        response = self.client.get(self.detail_url(collection2.id))
+        self.assertEqual(response.data['data_count'], 2)
+
+    def test_collection_status(self):
+        data_error = self._create_data(Data.STATUS_ERROR)
+        data_uploading = self._create_data(Data.STATUS_UPLOADING)
+        data_processing = self._create_data(Data.STATUS_PROCESSING)
+        data_waiting = self._create_data(Data.STATUS_WAITING)
+        data_resolving = self._create_data(Data.STATUS_RESOLVING)
+        data_done = self._create_data(Data.STATUS_DONE)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="error")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add(data_error, data_uploading, data_processing, data_waiting, data_resolving, data_done)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="uploading")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add(data_uploading, data_processing, data_waiting, data_resolving, data_done)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="processing")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add(data_processing, data_waiting, data_resolving, data_done)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="waiting")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add(data_waiting, data_resolving, data_done)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="resolving")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add(data_resolving, data_done)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="done")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add(data_done)
+
+        collection = Collection.objects.create(contributor=self.contributor, name="empty")
+        assign_perm("view_collection", AnonymousUser(), collection)
+        collection.data.add()
+
+        collections = self.client.get(self.list_url).data
+
+        get_collection = lambda collections, name: next(x for x in collections if x['name'] == name)
+        self.assertEqual(get_collection(collections, "error")['status'], Data.STATUS_ERROR)
+        self.assertEqual(get_collection(collections, "uploading")['status'], Data.STATUS_UPLOADING)
+        self.assertEqual(get_collection(collections, "processing")['status'], Data.STATUS_PROCESSING)
+        self.assertEqual(get_collection(collections, "waiting")['status'], Data.STATUS_WAITING)
+        self.assertEqual(get_collection(collections, "resolving")['status'], Data.STATUS_RESOLVING)
+        self.assertEqual(get_collection(collections, "done")['status'], Data.STATUS_DONE)
+        self.assertEqual(get_collection(collections, "empty")['status'], None)
 
 
 class ProcessTestCase(ResolweAPITestCase):